feat: restructure screens folder
