+++ lib/presentation/profile/pages/account.dart
@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:lecognition/common/widgets/appbar.dart';
+import 'package:lecognition/presentation/profile/pages/edit.dart';
 
 class AkunScreen extends StatelessWidget {
   const AkunScreen({super.key});
@@ -6,10 +8,7 @@ class AkunScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: const Text('Akun Pengguna'),
-        backgroundColor: Colors.amber,
-      ),
+      appBar: AppBarWidget(title: 'Akun Saya'),
       body: SingleChildScrollView(
         child: Padding(
           padding: const EdgeInsets.all(20.0),
@@ -20,9 +19,15 @@ class AkunScreen extends StatelessWidget {
               Center(
                 child: Column(
                   children: [
-                    const CircleAvatar(
-                      radius: 60,
-                      backgroundImage: AssetImage('assets/images/profile_pic.jpg'), // Ganti dengan gambar profil Anda
+                    Hero(
+                      tag: 'profile_image',
+                      child: CircleAvatar(
+                        radius: 60,
+                        child: Image.asset(
+                          'assets/images/icon.png',
+                          width: 100,
+                        ),
+                      ),
                     ),
                     const SizedBox(height: 20),
                     const Text(
@@ -34,7 +39,8 @@ class AkunScreen extends StatelessWidget {
                     ),
                     const SizedBox(height: 5),
                     const Text(
-                      'Frontend & Backend Developer',
+                      // 'Frontend & Backend Developer',
+                      '',
                       style: TextStyle(
                         fontSize: 16,
                         color: Colors.grey,
@@ -50,33 +56,42 @@ class AkunScreen extends StatelessWidget {
                 style: TextStyle(
                   fontWeight: FontWeight.bold,
                   fontSize: 20,
-                  color: Colors.teal,
+                  color: Colors.black,
                 ),
               ),
               const SizedBox(height: 20),
               _buildInfoCard(
                 icon: Icons.person,
-                title: 'Nama Lengkap',
+                title: 'Username',
                 subtitle: 'Nazwa Ayunda M',
+                context: context,
               ),
               _buildInfoCard(
                 icon: Icons.email,
                 title: 'Email',
                 subtitle: 'nazwam@example.com',
+                context: context,
               ),
-              _buildInfoCard(
-                icon: Icons.phone,
-                title: 'Nomor Telepon',
-                subtitle: '+6281234567890',
-              ),
+              // _buildInfoCard(
+              //   icon: Icons.phone,
+              //   title: 'Nomor Telepon',
+              //   subtitle: '+6281234567890',
+              // ),
               const SizedBox(height: 20),
               Center(
                 child: ElevatedButton.icon(
                   onPressed: () {
                     // Tambahkan aksi jika perlu
+                    Navigator.push(
+                      context,
+                      MaterialPageRoute(
+                        builder: (context) => EditAccount(),
+                      ),
+                    );
                   },
                   style: ElevatedButton.styleFrom(
-                    backgroundColor: Colors.teal,
+                    backgroundColor: Theme.of(context).colorScheme.primary,
+                    foregroundColor: Theme.of(context).colorScheme.onPrimary,
                     padding: const EdgeInsets.symmetric(horizontal: 30, vertical: 15),
                     shape: RoundedRectangleBorder(
                       borderRadius: BorderRadius.circular(30),
@@ -101,15 +116,19 @@ class AkunScreen extends StatelessWidget {
     required IconData icon,
     required String title,
     required String subtitle,
+    required BuildContext context,
   }) {
     return Card(
       elevation: 2,
       shape: RoundedRectangleBorder(
         borderRadius: BorderRadius.circular(15),
+        side: BorderSide(color: Colors.grey.shade200),
       ),
       margin: const EdgeInsets.symmetric(vertical: 10),
+      color: Colors.white,
+      shadowColor: Colors.grey,
       child: ListTile(
-        leading: Icon(icon, color: Colors.teal, size: 30),
+        leading: Icon(icon, color: Theme.of(context).colorScheme.primary, size: 30),
         title: Text(
           title,
           style: const TextStyle(fontWeight: FontWeight.bold, fontSize: 16),
